requirements.txt
pandas
python-dotenvgoogle-generativeai
flask-cors